==> notebooks/00_learn_lda.ipynb <==
# !python3 -m pip install gensim nltk pandas
# https://towardsdatascience.com/topic-modeling-with-latent-dirichlet-allocation-e7ff75290f8
# https://towardsdatascience.com/gaussian-mixture-model-clusterization-how-to-select-the-number-of-components-clusters-553bef45f6e4


# Topic extraction using LDA

## Importing modules & defining globals


import os
import numpy as np
import pandas as pd
from typing import Union

np.random.seed(42)

DATA_PATH = "../data/raw"


## Loading data


io = os.path.join(DATA_PATH, "Pubmed5k.xlsx")
sheet_name = "random 5k"
df = pd.read_excel(io, sheet_name, index_col=[0])


## Exploring the data


df.head()


TODO: add more EDA


## Following simple approach from [here](https://towardsdatascience.com/topic-modeling-and-latent-dirichlet-allocation-in-python-9bf156893c24)

### Importing required packages & downloading assets


import nltk

nltk.download("wordnet")
nltk.download("stopwords")


from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import STOPWORDS
from gensim.corpora import Dictionary
from gensim.models import TfidfModel, LdaMulticore
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer


### combining `gensim`'s stopwords w/ `nltk`'s


STOPWORDS = STOPWORDS.union(set(stopwords.words("english")))


### Preprocessing


# combine article title w/ abstract to obtain a single document
df_joined = df["Title"] + " " + df["Abstract"]
df_joined.name = "document"


def preprocess(
    document: str,
    min_len: int = 2,
    max_len: int = 15,
    stopwords: frozenset = frozenset(),
    pos: Union[str, list] = "n",
) -> str:
    """Tokenise the document, drop stopwords, then lowercasing and lemmatising each token

    TODO:
    """
    pp_doc = simple_preprocess(document, min_len=min_len, max_len=max_len)
    doc_non_stop = [token for token in pp_doc if token not in stopwords]
    lemmatise = WordNetLemmatizer().lemmatize
    if type(pos) == str:
        return [lemmatise(token, pos=pos) for token in doc_non_stop]
    else:
        result = doc_non_stop
        for p in pos:
            result = [lemmatise(token, pos=p) for token in result]
        return result


# preprocessing the corpus
pos = ["a", "n", "r", "s", "v"]
df_processed = df_joined.apply(preprocess, stopwords=STOPWORDS, pos=pos)


# show a random document
" ".join(df_processed.iloc[np.random.randint(0, len(df_processed))])


### Counting tokens


dictionary = Dictionary(df_processed)


### Filtering out words


# removing tokens that appear in less that 1% (.01) or in more than 50% (.5) of the corpus
dictionary.filter_extremes(int(0.01 * len(df_processed)), 0.5, None)


### Creating BoW


df_bow = df_processed.apply(dictionary.doc2bow)


### Creating TF-IDF model


tfidf_model = TfidfModel(df_bow.values.tolist())


df_tfidf = df_bow.apply(lambda doc: tfidf_model[doc])


### LDA model


lda_model = LdaMulticore(df_bow, id2word=dictionary)
lda_model.print_topics()


def get_top_n_topics(
    corpus: pd.DataFrame, model: LdaMulticore, n: int = 3
) -> pd.DataFrame:
    """calculates the top `n` topics for each document in `df` through the model
    
    Parameters:
    -----------
    
    Returns:
    --------
    
    TODO: complete the pydoc
    """
    # if not isinstance(df, pd.DataFrame):
    #     df = pd.DataFrame(df)
    df_lda = corpus.apply(
        lambda doc: sorted(model[doc], key=(lambda tup: tup[1]), reverse=True)[:n]
    )
    corpus = pd.DataFrame(corpus)

    for i in range(n):
        corpus[f"topic_{i+1}"] = df_lda.apply(
            lambda topics: topics[i][0] if len(topics) > i else None
        )
        corpus[f"topic_{i+1}_prop"] = df_lda.apply(
            lambda topics: topics[i][1] if len(topics) > i else None
        )

    # FIXME: some documents might have less than `n` possible topics
    # to avoid N/As, set them to the previous topic
    for i in range(2, n + 1):
        idx = corpus[f"topic_{i}"].isna()
        cols = [f"topic_{i}", f"topic_{i}_prop"]
        cols_prev = [f"topic_{i-1}", f"topic_{i-1}_prop"]
        corpus.loc[idx, cols] = corpus.loc[idx, cols_prev].values

    return corpus.drop(columns=["document"])


df_topic = get_top_n_topics(df_bow, lda_model, 3)
df_topic.head()


df_topic.to_csv(os.path.join(DATA_PATH, "bow_100_topic.csv"))


lda_model = LdaMulticore(df_tfidf, id2word=dictionary)
lda_model.print_topics()


df_topic = get_top_n_topics(df_tfidf, lda_model, 3)
df_topic.head()


df_topic.to_csv(os.path.join(DATA_PATH, "tfidf_100_topic.csv"))
