==> notebooks/01_feature_engineering.ipynb <==
# !python3 -m pip install langdetect


# Feature Engineering of corpus

The following resources were found helpful devising a strategy for feature engineering:

- [Text Analysis & Feature Engineering with NLP](https://towardsdatascience.com/text-analysis-feature-engineering-with-nlp-502d6ea9225d)
- [Topic Modelling and Latent Dirichlet Allocation \(LDA\) in Python](https://towardsdatascience.com/topic-modeling-and-latent-dirichlet-allocation-in-python-9bf156893c24)

## Importing modules &amp; defining globals


import os
import numpy as np
import pandas as pd

from gensim.parsing.preprocessing import STOPWORDS
from nltk.corpus import stopwords

cwd = os.getcwd()
os.chdir("../")
from lda_utils import preprocess

os.chdir(cwd)
# from gensim.utils import simple_preprocess
# from nltk.stem import WordNetLemmatizer

# from typing import Union

np.random.seed(42)

DATA_PATH = "../data/raw"
STOPWORDS = STOPWORDS | frozenset(stopwords.words("english"))


## Loading dataset


io = "Pubmed5k.xlsx"
sheet_name = "random 5k"
df = pd.read_excel(os.path.join(DATA_PATH, io), sheet_name=sheet_name, index_col=[0])
df.head()


## Exploring dataset


df_lower = df["Abstract"].str.lower()

# some records have no clear abstract
idx_no_abs = df_lower.str.find("no abstract") != -1

df_processed = df[~idx_no_abs]
# df_no_abs = df[idx_no_abs]
# df_no_abs.loc[:, "Abstract"] = ""
# df_processed.loc[idx_no_abs, "Abstract"] = ""

# df_no_abstract = df[idx_no_abs]
# df_corpus = df[~idx_no_abs]


df_len = df_processed.apply({"Title": len, "Abstract": len})
df_len.describe()


some of the statistics for length of the documents are illogical.
e.g: min(Abstract) = 1


df_len["Abstract"].value_counts().sort_index().head(10)


df_processed[df_len["Abstract"] <= 104]["Abstract"]


df_processed[df_len["Abstract"] == 43]["Title"].values


df_len["Title"].value_counts().sort_index().head(10)


df_processed[df_len["Title"] <= 31]["Title"]


df_altered = df[idx_no_abs]
df_altered = pd.concat([df_altered, df_processed[df_len["Abstract"] <= 43]])
df_altered.loc[:, "Abstract"] = ""
df_altered.shape


df_altered.head()


# df_processed.drop(index=df_altered.index)
df_processed = df.drop(index=df_altered.index)
df_processed = pd.concat([df_processed, df_altered])
df_processed.shape


## Preprocessing

> Note: the approaches here are for a unigram model.

after exploring the dataset, and slightly tweaking some features, the two columns of the set are joined to make a single feature (document per record) to be able to operate on the dataset


corpus = df_processed["Title"] + r" " + df_processed["Abstract"]
corpus.name = "document"
corpus.head()


Defining helper function. Basically, bundling the preprocess subroutine into a function

For each document:

1. Tokenising the document.
2. lowercasing the tokens.
3. lemmatising the tokens.
4. dropping stop words.


- Calling the preprocessing subroutine on the dataset


pos = ["a", "n", "r", "s", "v"]
corpus_processed = corpus.apply(preprocess, stopwords=STOPWORDS, pos=pos)
corpus_processed.head()


## Saving the final feature engineered dataset


corpus_processed.to_csv(os.path.join(DATA_PATH, "corpus_clean.csv"))
