==> notebooks/03_unigram_model.ipynb <==
# Training a Unigram `LdaMulticore` model

## Importing modules &amp; defining globals


import os
import pickle
import numpy as np
import pandas as pd

from ast import literal_eval
from gensim.models import LdaMulticore, CoherenceModel

cwd = os.getcwd()
os.chdir("..")
from lda_utils import get_top_n_topics

os.chdir(cwd)

RAW_DATA_PATH = "../data/raw"
PKL_DATA_PATH = "../data/pickles"


## Loading preprocessed dataset


corpus = pd.read_csv(
    os.path.join(RAW_DATA_PATH, "corpus_clean.csv"), index_col=["ArticleID"]
)
corpus = corpus["document"].apply(literal_eval)
corpus.head()


## Loading hyperparameters


h_params = np.load(os.path.join(RAW_DATA_PATH, "hyperparameters.npy"))
k = int(h_params[0])
a = h_params[1]
k, a


## Load arguments required by the model

The arguments are dictionary of id/word, and the bow corpus


params = {}
for _, _, filenames in os.walk(PKL_DATA_PATH):
    for fname in filenames:
        if not fname.endswith(".pkl"):
            continue
        path = os.path.join(PKL_DATA_PATH, fname)
        file_ref = open(path, "rb")
        params[fname.replace(".pkl", "")] = pickle.load(file_ref)
        file_ref.close()


## Building a unigram model


lda_model = LdaMulticore(
    params["bow"],
    num_topics=k,
    id2word=params["id2word"],
    eta="auto",
    alpha=a,
    random_state=42,
)


## Assigning topics

now we use the model to assign topics to records/docs/articles


n = 3

df_topics = get_top_n_topics(params["bow"], lda_model, n)
df_topics.head()


## Saving findings


df_topics.to_csv(os.path.join(RAW_DATA_PATH, "topics.csv"))


## Evaluating model


co_model = CoherenceModel(
    lda_model, texts=corpus, dictionary=params["id2word"], coherence="u_mass"
)
co_model.get_coherence()


## Saving the model


path = os.path.join(PKL_DATA_PATH, "lda_model.pkl")
lda_model.save(path)
